+++ axelor-base/src/main/java/com/axelor/apps/base/db/IImports.java
@@ -21,6 +21,7 @@ package com.axelor.apps.base.db;
  * Interface of Imports package. Enum all static variable of packages.
  * 
  */
+@Deprecated
 public interface IImports {
 
 	/**