+++ axelor-base/src/main/java/com/axelor/apps/base/service/PartnerService.java
@@ -17,6 +17,15 @@
  */
 package com.axelor.apps.base.service;
 
+import java.time.LocalDate;
+import java.util.ArrayList;
+import java.util.HashMap;
+import java.util.List;
+import java.util.Map;
+import java.util.Set;
+
+import javax.inject.Singleton;
+
 import com.axelor.apps.base.db.Address;
 import com.axelor.apps.base.db.BankDetails;
 import com.axelor.apps.base.db.Currency;
@@ -39,15 +48,6 @@ import com.axelor.inject.Beans;
 import com.google.common.base.Strings;
 import com.google.inject.Inject;
 import com.google.inject.persist.Transactional;
-
-import java.time.LocalDate;
-import java.util.ArrayList;
-import java.util.HashMap;
-import java.util.List;
-import java.util.Map;
-import java.util.Set;
-
-import javax.inject.Singleton;
 
 @Singleton
 public class PartnerService {
@@ -367,4 +367,23 @@ public class PartnerService {
 			return null;
 		}
 	}
+	
+	/**
+	 * Get the partner language code. If null, return the default partner language.
+	 * @param partner
+	 * @return
+	 */
+	public String getPartnerLanguageCode(Partner partner)  {
+		
+		String locale = null;
+
+    	if(partner != null && partner.getLanguage() != null)  {
+    		locale = partner.getLanguage().getCode();
+    	}
+    	if (!Strings.isNullOrEmpty(locale))  {  return locale;  }
+    	
+    	return Beans.get(AppBaseService.class).getDefaultPartnerLanguageCode();
+		
+	}
+	
 }