+++ axelor-base/src/main/resources/sql/unitConversionMigration.sql
@@ -1,3 +1,0 @@
-UPDATE base_unit_conversion
-SET type_select = 2
-WHERE type_select IS NULL