+++ axelor-crm/src/main/java/com/axelor/apps/crm/db/ICrmBatch.java
@@ -23,6 +23,7 @@ package com.axelor.apps.crm.db;
  * @author dubaux
  * 
  */
+@Deprecated
 public interface ICrmBatch {