+++ axelor-production/src/main/java/com/axelor/apps/production/service/MrpServiceProductionImpl.java
@@ -150,6 +150,14 @@ public class MrpServiceProductionImpl extends MrpServiceImpl  {
  					Product product = prodProduct.getProduct();
 					
 					if(this.isMrpProduct(product))  {
+
+						// A component of a manuf order that is not loaded on MRP because there is no default BOM or 
+						// because the component of manuf order is not a component of the bill of material, we add it with the level of manuf order product + 1.
+						if(!this.productMap.containsKey(product.getId()))  {
+							this.assignProductAndLevel(product, manufOrder.getProduct());
+							this.createAvailableStockMrpLine(product, manufOrder.getProdProcess().getStockLocation());
+						}
+
 						mrp.addMrpLineListItem(this.createMrpLine(product, manufOrderNeedMrpLineType, prodProduct.getQty(), 
 								manufOrder.getPlannedStartDateT().toLocalDate(), BigDecimal.ZERO, stockLocation, manufOrder));
 					}
@@ -272,7 +280,25 @@ public class MrpServiceProductionImpl extends MrpServiceImpl  {
 	}
 	
 	
+	/**
+	 * Add a component product of a manuf order where the component product is not contained on the default bill of material of the produced product.
+	 * @param manufOrderComponentProduct
+	 * @param manufOrderProducedProduct
+	 */
+	protected void assignProductAndLevel(Product manufOrderComponentProduct, Product manufOrderProducedProduct)  {
+		
+		log.debug("Add of the product : {}", manufOrderComponentProduct.getFullName());
+		this.productMap.put(manufOrderComponentProduct.getId(),  this.getMaxLevel(manufOrderProducedProduct, 0)+1);
+		
+	}
 	
+	protected void createAvailableStockMrpLine(Product product, StockLocation stockLocation) throws AxelorException  {
+		
+		MrpLineType availableStockMrpLineType = this.getMrpLineType(MrpLineTypeRepository.ELEMENT_AVAILABLE_STOCK);
+		
+		mrp.addMrpLineListItem(this.createAvailableStockMrpLine(productRepository.find(product.getId()), stockLocation, availableStockMrpLineType));
+				
+	}
 	
 }