+++ axelor-production/src/main/java/com/axelor/apps/production/service/OperationOrderWorkflowService.java
@@ -23,10 +23,13 @@ import java.time.Duration;
 import java.time.LocalDateTime;
 import java.util.List;
 
+import org.apache.commons.collections.CollectionUtils;
+
 import com.axelor.app.production.db.IManufOrder;
 import com.axelor.app.production.db.IOperationOrder;
 import com.axelor.app.production.db.IWorkCenter;
 import com.axelor.apps.production.db.Machine;
+import com.axelor.apps.production.db.ManufOrder;
 import com.axelor.apps.production.db.OperationOrder;
 import com.axelor.apps.production.db.OperationOrderDuration;
 import com.axelor.apps.production.db.ProdHumanResource;
@@ -38,6 +41,7 @@ import com.axelor.apps.production.db.repo.ProductionConfigRepository;
 import com.axelor.apps.production.service.app.AppProductionService;
 import com.axelor.apps.production.service.config.ProductionConfigService;
 import com.axelor.auth.AuthUtils;
+import com.axelor.db.JPA;
 import com.axelor.exception.AxelorException;
 import com.axelor.inject.Beans;
 import com.google.inject.Inject;
@@ -71,7 +75,9 @@ public class OperationOrderWorkflowService {
 	@Transactional(rollbackOn = {AxelorException.class, Exception.class})
 	public OperationOrder plan(OperationOrder operationOrder) throws AxelorException  {
 
-		Beans.get(OperationOrderService.class).createToConsumeProdProductList(operationOrder);
+		if (CollectionUtils.isEmpty(operationOrder.getToConsumeProdProductList())) {
+			Beans.get(OperationOrderService.class).createToConsumeProdProductList(operationOrder);
+		}
 
 		operationOrder.setPlannedStartDateT(this.getLastOperationOrder(operationOrder));
 
@@ -82,7 +88,10 @@ public class OperationOrderWorkflowService {
 						Duration.between(operationOrder.getPlannedStartDateT(), operationOrder.getPlannedEndDateT())
 				));
 
-		operationOrderStockMoveService.createToConsumeStockMove(operationOrder);
+		ManufOrder manufOrder = operationOrder.getManufOrder();
+		if (manufOrder == null || manufOrder.getIsConsProOnOperation()) {
+			operationOrderStockMoveService.createToConsumeStockMove(operationOrder);
+		}
 
 		operationOrder.setStatusSelect(IOperationOrder.STATUS_PLANNED);
 
@@ -251,6 +260,9 @@ public class OperationOrderWorkflowService {
 		if (oldStatus == IOperationOrder.STATUS_IN_PROGRESS) {
 			stopOperationOrderDuration(operationOrder);
 		}
+		if (operationOrder.getConsumedStockMoveLineList() != null) {
+			operationOrder.getConsumedStockMoveLineList().forEach(stockMoveLine -> stockMoveLine.setConsumedOperationOrder(null));
+		}
 		operationOrderStockMoveService.cancel(operationOrder);
 
 		operationOrderRepo.save(operationOrder);
@@ -303,8 +315,13 @@ public class OperationOrderWorkflowService {
 		Duration totalDuration = Duration.ZERO;
 
 		List<OperationOrderDuration> operationOrderDurations = operationOrder.getOperationOrderDurationList();
-		for (OperationOrderDuration operationOrderDuration : operationOrderDurations) {
-			totalDuration = totalDuration.plus(Duration.between(operationOrderDuration.getStartingDateTime(), operationOrderDuration.getStoppingDateTime()));
+		if (operationOrderDurations != null) {
+			for (OperationOrderDuration operationOrderDuration : operationOrderDurations) {
+			    if (operationOrderDuration.getStartingDateTime() != null
+						&& operationOrderDuration.getStoppingDateTime() != null) {
+					totalDuration = totalDuration.plus(Duration.between(operationOrderDuration.getStartingDateTime(), operationOrderDuration.getStoppingDateTime()));
+				}
+			}
 		}
 
 		return totalDuration;