+++ axelor-project/src/main/java/com/axelor/apps/project/db/repo/ProjectManagementRepository.java
@@ -42,7 +42,7 @@ public class ProjectManagementRepository extends ProjectRepository {
 	}
 
 	public static void setAllProjectMembersUserSet(Project project) {
-		if (project.getParentProject() == null) {
+		if (project.getParentProject() == null && project.getChildProjectList() != null) {
 			project.getChildProjectList().stream()
 					.filter(Project::getExtendsMembersFromParent)
                     .peek(p -> project.getMembersUserSet().forEach(p::addMembersUserSetItem))