+++ axelor-supplychain/src/main/java/com/axelor/apps/supplychain/web/SaleOrderController.java
@@ -50,7 +50,6 @@ import com.axelor.apps.supplychain.service.SaleOrderServiceSupplychainImpl;
 import com.axelor.apps.supplychain.service.SaleOrderStockService;
 import com.axelor.apps.supplychain.service.app.AppSupplychainService;
 import com.axelor.db.JPA;
-import com.axelor.exception.AxelorException;
 import com.axelor.exception.service.TraceBackService;
 import com.axelor.i18n.I18n;
 import com.axelor.inject.Beans;
@@ -67,7 +66,9 @@ import com.google.inject.Singleton;
 
 @Singleton
 public class SaleOrderController{
-	
+
+	private final String SO_LINES_WIZARD_QTY_TO_INVOICE_FIELD = "qtyToInvoice";
+
 	@Inject
 	private SaleOrderServiceSupplychainImpl saleOrderServiceSupplychain;
 
@@ -210,9 +211,9 @@ public class SaleOrderController{
 			saleOrderLineListContext = (List<Map<String,Object>>)
 					request.getRawContext().get("saleOrderLineList");
 			for (Map<String, Object> map : saleOrderLineListContext ) {
-				if (map.get("amountToInvoice") != null) {
+				if (map.get(SO_LINES_WIZARD_QTY_TO_INVOICE_FIELD) != null) {
 					BigDecimal qtyToInvoiceItem = new BigDecimal(
-							map.get("amountToInvoice").toString()
+							map.get(SO_LINES_WIZARD_QTY_TO_INVOICE_FIELD).toString()
 					);
 					if (qtyToInvoiceItem.compareTo(BigDecimal.ZERO) != 0) {
 						Long SOlineId = new Long((Integer) map.get("id"));
@@ -470,8 +471,8 @@ public class SaleOrderController{
 						.context("_showRecord", String.valueOf(saleOrder.getId())).map());
 				response.setCanClose(true);
 			}
-		}catch(AxelorException ae){
-			response.setFlash(ae.getLocalizedMessage());
+		}catch(Exception e){
+			response.setFlash(e.getLocalizedMessage());
 		}
 	}